@@ -3,24 +3,33 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.3.0",
-    "@fortawesome/free-brands-svg-icons": "^6.0.0",
-    "@fortawesome/free-solid-svg-icons": "^6.0.0",
-    "@fortawesome/react-fontawesome": "^0.1.17",
-    "moment": "^2.29.3",
-    "next": "11.1.3",
-    "next-seo": "^5.4.0",
-    "react": "^17.0.2",
-    "react-accessible-accordion": "^4.0.0",
-    "react-big-calendar": "^0.40.0",
-    "react-countdown": "^2.3.2",
-    "react-dom": "^17.0.2",
-    "sass": "^1.49.9"
+    "@fortawesome/fontawesome-svg-core": "^6.4.2",
+    "@fortawesome/free-brands-svg-icons": "^6.4.2",
+    "@fortawesome/free-regular-svg-icons": "^6.5.1",
+    "@fortawesome/free-solid-svg-icons": "^6.5.1",
+    "@fortawesome/react-fontawesome": "^0.2.1",
+    "@netlify/plugin-nextjs": "^5.9.4",
+    "@testing-library/dom": "^10.4.0",
+    "a": "^3.0.1",
+    "dotenv": "^16.4.5",
+    "googleapis": "^144.0.0",
+    "moment": "^2.29.4",
+    "next": "^15.1.6",
+    "next-seo": "^6.5.0",
+    "react": "^18.3.1",
+    "react-accessible-accordion": "^5.0.0",
+    "react-big-calendar": "^1.8.4",
+    "react-countdown": "^2.3.5",
+    "react-dom": "^18.3.1",
+    "react-social-media-embed": "^2.5.18",
+    "sass": "^1.84.0",
+    "three": "^0.157.0",
+    "yarn": "^1.22.21"
   },
   "scripts": {
     "start": "next dev",
     "start-production": "next start",
-    "build": "next build",
+    "build": "npm run alp && npm run ofp && npm run nep && npm run thp && npm run hep && npm run gmp && next build",
     "lint-js": "eslint \"**/*.js\"",
     "lint-js-fix": "eslint --fix \"**/*.js\"",
     "lint-css": "stylelint \"**/*.css\" \"**/*.scss\"",
@@ -29,6 +38,12 @@
     "lint-fix": "eslint --fix \"**/*.js\" && stylelint --fix \"**/*.css\" \"**/*.scss\"",
     "reg": "node scripts/recurring-event-generator.mjs",
     "seg": "node scripts/single-event-generator.mjs",
+    "alp": "node scripts/alumni-parser.mjs",
+    "ofp": "node scripts/officer-parser.mjs",
+    "nep": "node scripts/news-parser.mjs",
+    "thp": "node scripts/town-hall-generator.mjs",
+    "gmp": "node scripts/gm-generator.mjs",
+    "hep": "node scripts/history-events.mjs",
     "test": "jest --env=jsdom"
   },
   "eslintConfig": {
@@ -58,21 +73,23 @@
     ]
   },
   "devDependencies": {
-    "@axe-core/react": "^4.4.2",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.5",
+    "@axe-core/react": "^4.7.3",
+    "@testing-library/jest-dom": "^6.1.3",
+    "@testing-library/react": "^16.2.0",
     "csv-parser": "^3.0.0",
     "eslint": "^7.32.0",
-    "eslint-config-next": "^11.1.3",
+    "eslint-config-next": "^13.4.2",
     "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-react": "^7.29.4",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.33.2",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^27.5.1",
-    "jest-axe": "^5.0.1",
-    "stylelint": "^13.13.1",
-    "stylelint-config-recommended": "^5.0.0",
-    "stylelint-config-sass-guidelines": "^8.0.0",
-    "stylelint-scss": "^3.21.0"
+    "jest": "^29.7.0",
+    "jest-axe": "^9.0.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "postcss": "^8.4.31",
+    "stylelint": "^16.14.1",
+    "stylelint-config-recommended": "^15.0.0",
+    "stylelint-config-sass-guidelines": "^12.1.0",
+    "stylelint-scss": "^6.11.0"
   }
 }
