@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import { useRouter } from 'next/router';
-import React, { Children } from 'react';
+import { Children, cloneElement } from 'react';
 
 const ActiveLink = ({ children, activeClassName, ...props }) => {
   const { asPath } = useRouter();
@@ -8,18 +8,19 @@ const ActiveLink = ({ children, activeClassName, ...props }) => {
   const childClassName = child.props.className || '';
 
   // if path equals href append activeClassName to childClassName otherwise return childClassName
-  const className = asPath === props.href || asPath === props.as
-    ? `${childClassName} ${activeClassName}`.trim()
-    : childClassName;
+  const className =
+    asPath === props.href || asPath === props.as
+      ? `${childClassName} ${activeClassName}`.trim()
+      : childClassName;
 
-    return (
-      //clones child with className if className exists
-      <Link {...props}>
-        {React.cloneElement(child, {
-          className: className || null,
-        })}
-      </Link>
-    );
+  return (
+    //clones child with className if className exists
+    (<Link {...props} legacyBehavior>
+      {cloneElement(child, {
+        className: className || null,
+      })}
+    </Link>)
+  );
 };
 
 export default ActiveLink;
