@@ -1,24 +1,23 @@
-import Image from 'next/image';
-import React from 'react';
+import Image from 'next/legacy/image';
 
-function Officer(props){
-	return (
-		<div className="officer-grid-row">
-			<div className="officer-grid-col">
-				<Image
-					className="officer-image"
-					src={props.officer.image}
-					alt={props.officer.name}
-					width={70}
-					height={70}
-				/>
-			</div>
-			<div className="officer-grid-col officer-info">
-				<h3 className="officer-title">{props.officer.name}</h3>
-				<p>{props.officer.title}</p>
-			</div>
-		</div>
-	);
+function Officer(props) {
+  return (
+    <div className="officer-grid-row">
+      <div className="officer-grid-col">
+        <Image
+          className="officer-image"
+          src={props.officer.image}
+          alt={props.officer.name}
+          width={70}
+          height={70}
+        />
+      </div>
+      <div className="officer-grid-col officer-info">
+        <h3 className="officer-title">{props.officer.name}</h3>
+        <p>{props.officer.title}</p>
+      </div>
+    </div>
+  );
 }
 
 function Officers(props){
