@@ -1,16 +1,23 @@
-import Image from 'next/image';
-import React from 'react';
+import Image from 'next/legacy/image';
 
 function TechGalaBanner() {
-	return (
-        <div>
-            <Image src={'/images/techgala/techgala-banner-dark.jpg'} alt="2021 tech gala banner" height={410} width={780}/>
-            <div style={{ marginTop: '30px' }}>
-                <h1>Tech Gala</h1>
-                <p>Here are the wonderful projects showcased at our Winter 2021 Tech Gala from students of UCLA!</p>
-            </div>
-        </div>
-	);
+  return (
+    <div>
+      <Image
+        src={'/images/techgala/techgala-banner-dark.jpg'}
+        alt="2021 tech gala banner"
+        height={410}
+        width={780}
+      />
+      <div style={{ marginTop: '30px' }}>
+        <h1>Tech Gala</h1>
+        <p>
+          Here are the wonderful projects showcased at our Winter 2021 Tech Gala
+          from students of UCLA!
+        </p>
+      </div>
+    </div>
+  );
 }
 
 export default TechGalaBanner;
