@@ -1,44 +1,48 @@
 export default {
-	images: [
-		'/images/carousel/01.jpg',
-		'/images/carousel/02.jpg',
-		'/images/carousel/03.jpg',
-		'/images/carousel/04.jpg',
-		'/images/carousel/05.jpg',
-		'/images/carousel/06.jpg',
-		'/images/carousel/07.jpg',
-		'/images/carousel/08.jpg',
-		'/images/carousel/09.jpg',
-		'/images/carousel/10.jpg',
-		'/images/carousel/11.jpg',
-		'/images/carousel/12.jpg',
-		'/images/carousel/13.jpg',
-		'/images/carousel/14.jpg',
-		'/images/carousel/15.jpg',
-		'/images/carousel/16.jpg',
-		'/images/carousel/17.jpg',
-		'/images/carousel/18.jpg',
-		'/images/carousel/19.jpg',
-		'/images/carousel/20.jpg',
-		'/images/carousel/21.jpg',
-		'/images/carousel/22.jpg',
-		'/images/carousel/23.jpg',
-		'/images/carousel/24.jpg',
-		'/images/carousel/25.jpg',
-		'/images/carousel/26.jpg',
-		'/images/carousel/27.jpg',
-		'/images/carousel/28.jpg',
-		'/images/carousel/29.jpg',
-		'/images/carousel/30.jpg',
-		'/images/carousel/31.jpg',
-		'/images/carousel/32.jpg',
-		'/images/carousel/33.jpg',
-		'/images/carousel/34.jpg',
-		'/images/carousel/35.jpg',
-		'/images/carousel/36.jpg',
-		'/images/carousel/37.jpg',
-		'/images/carousel/38.jpg',
-		'/images/carousel/39.jpg',
-		'/images/carousel/40.jpg',
-	],
+  images: [
+    '/images/carousel/41.jpg',
+    '/images/carousel/42.jpg',
+    '/images/carousel/43.jpg',
+    '/images/carousel/44.jpg',
+    '/images/carousel/01.jpg',
+    '/images/carousel/02.jpg',
+    '/images/carousel/03.jpg',
+    '/images/carousel/04.jpg',
+    '/images/carousel/05.jpg',
+    '/images/carousel/06.jpg',
+    '/images/carousel/07.jpg',
+    '/images/carousel/08.jpg',
+    '/images/carousel/09.jpg',
+    '/images/carousel/10.jpg',
+    '/images/carousel/11.jpg',
+    '/images/carousel/12.jpg',
+    '/images/carousel/13.jpg',
+    '/images/carousel/14.jpg',
+    '/images/carousel/15.jpg',
+    '/images/carousel/16.jpg',
+    '/images/carousel/17.jpg',
+    '/images/carousel/18.jpg',
+    '/images/carousel/19.jpg',
+    '/images/carousel/20.jpg',
+    '/images/carousel/21.jpg',
+    '/images/carousel/22.jpg',
+    '/images/carousel/23.jpg',
+    '/images/carousel/24.jpg',
+    '/images/carousel/25.jpg',
+    '/images/carousel/26.jpg',
+    '/images/carousel/27.jpg',
+    '/images/carousel/28.jpg',
+    '/images/carousel/29.jpg',
+    '/images/carousel/30.jpg',
+    '/images/carousel/31.jpg',
+    '/images/carousel/32.jpg',
+    '/images/carousel/33.jpg',
+    '/images/carousel/34.jpg',
+    '/images/carousel/35.jpg',
+    '/images/carousel/36.jpg',
+    '/images/carousel/37.jpg',
+    '/images/carousel/38.jpg',
+    '/images/carousel/39.jpg',
+    '/images/carousel/40.jpg',
+  ],
 };
