@@ -1,8 +1,50 @@
+import brief from '../public/images/impact/events/brief.png';
+import careerfair from '../public/images/impact/events/careerfair.png';
 import careersSocialImpactTech from '../public/images/impact/events/careers-social-tech.jpg';
+import h2i from '../public/images/impact/events/h2i2024.jpg';
 import introTechEthics from '../public/images/impact/events/intro-tech-ethics.png';
+import pitchathon from '../public/images/impact/events/pitchathon.png';
 import policyPitchathon21 from '../public/images/impact/events/policy-pitchathon-2021.png';
+import socialpanel from '../public/images/impact/events/socialpanel.jpeg';
 
 export const pastEvents = [
+
+  {
+    title: 'Social Good Pitch-a-Thon',
+    quarter: 'Fall 2023',
+    img: pitchathon,
+    alt: 'A graphic for the Social Good Pitch-a-Thon',
+    description: 'A hackathon-like event where participants teamed up to pitch technical solutions to real-world social justice and DEI issues, build connections with mentors, and win prizes.',
+  },
+  {
+    title: 'Hack2Impact',
+    quarter: 'Every Spring',
+    img: h2i,
+    alt: 'Over 20 people stand in front of a projected screen that says "Hack2Impact 2024"',
+    description: 'Our flagship event in collaboration with LA Blueprint — an annual social good hackathon that challenges participants to develop technical solutions for local nonprofit organizations\' needs.',
+  },
+  {
+    title: 'How to Get Your Start in AI Policy at UCLA',
+    quarter: 'Coming soon!',
+    img: brief,
+    alt: 'A robotic arm holding a scale',
+    description: 'A resource created alongside the AI Robotics Ethics Society and AI Safety at UCLA — get the inside scoop on how to get started in the field of AI policy as a UCLA undergraduate.',
+    // link: 'https://docs.google.com/document/d/15hA2d5xXQifDabpwsogzKdfrXy8SCtPm2IY4QqqOO0w/edit?usp=sharing',
+  },
+  {
+    title: 'Tech for Good Career Fair',
+    quarter: 'Winter 2024',
+    img: careerfair,
+    alt: 'A graphic for the Tech for Good Career Fair',
+    description: 'In collaboration with UPE Advocacy — an annual career fair that allows attendees to interact with companies, organizations, and opportunities that utilize technology for positive social impact.',
+  },
+  {
+    title: 'Social Impact Panel',
+    quarter: 'Winter 2024',
+    img: socialpanel,
+    alt: 'A laptop showing the panelists and moderators sits on a podium in a dimly lit room',
+    description: 'In collaboration with LA Blueprint — our audience heard from 4 industry professionals working on high-impact tech initiatives and gained insights into aligning their skills and passions with a career that makes a difference.',
+  },
   {
     title: 'Careers in Social Impact Tech',
     quarter: 'Winter 2021',
@@ -31,103 +73,40 @@ export const pastEvents = [
 
 export const people = [
   {
-    name: 'Maya Raman',
-    pronouns: 'she/they',
+    name: 'Naisha Agarwal',
+    pronouns: 'she/her',
     position: 'Impact Director',
     major: 'Computer Science Major',
-    year: 2023,
-    img: '/images/officers/MayaRaman.jpg',
-    alt: 'Maya smiling surrounded by plants',
+    year: 2026,
+    img: '/images/officers/NaishaAgarwal.jpg',
+    alt: 'Naisha similing in a car',
   },
   {
-    name: 'Alyssa Wang',
+    name: 'Aarohi Vohra',
     pronouns: 'she/her',
-    position: 'Impact Officer',
-    major: 'Computer Science Major',
-    year: 2022,
-    img: '/images/officers/AlyssaWang.jpeg',
+    position: 'Impact Director',
+    major: 'Math Major',
+    year: 2026,
+    img: '/images/officers/AarohiVohra.jpg',
     alt: 'Alyssa smiling in front of a courtyard',
   },
   {
-    name: 'Rohan Bansal',
-    pronouns: 'he/him',
-    position: 'Impact Officer',
-    major: 'Math of Computation Major',
-    year: 2022,
-    img: '/images/officers/RohanBansal.jpg',
+    name: 'Anne Do',
+    pronouns: 'she/her',
+    position: 'Impact Intern',
+    major: 'Computer Science Major',
+    year: 2027,
+    img: '/images/officers/AnneDo.jpg',
     alt: 'Rohan posing on a bridge',
   },
   {
-    name: 'Jason Jewik',
-    pronouns: 'he/him',
-    position: 'Impact Officer',
+    name: 'Elena Yu',
+    pronouns: 'she/her',
+    position: 'Impact Intern',
     major: 'Computer Science Major',
-    year: 2022,
-    img: '/images/officers/JasonJewik.jpg',
-    alt: 'a photo of jason smiling in front of some plants',
-  },
-  {
-    name: 'Hannah Chu',
-    pronouns: 'she/her',
-    position: 'Impact Officer',
-    major: 'Linguistics Major',
-    year: 2024,
-    img: '/images/officers/HannahChu.jpg',
-    alt: 'Hannah smiling on top of a hill',
-  },
-  {
-    name: 'Nisha McNealis',
-    pronouns: 'she/her',
-    position: 'Impact Officer',
-    major: 'Computer Science Major',
-    year: 2023,
-    img: '/images/officers/NishaMcNealis.jpeg',
-    alt: 'Nisha smiling in front of a wood backing',
-  },
-  {
-    name: 'Mia Luong',
-    pronouns: 'she/her',
-    position: 'Impact Intern',
-    major: 'Statistics Major',
-    year: 2024,
-    img: '/images/officers/MiaLuong.jpg',
-    alt: 'Mia posing in front of hills and mountains',
-  },
-  {
-    name: 'Agrim Gupta',
-    pronouns: 'he/him',
-    position: 'Impact Intern',
-    major: 'Linguistics & Computer Science + Statistics Major',
-    year: 2023,
-    img: '/images/officers/AgrimGupta.jpeg',
-    alt: 'Agrim smiling in front of art',
-  },
-  {
-    name: 'Diya Beliga',
-    pronouns: 'she/her',
-    position: 'Impact Intern',
-    major: 'Computer Science and Engineering Major',
-    year: 2024,
-    img: '/images/officers/NoImage.jpeg',
-    alt: '',
-  },
-  {
-    name: 'Brendan Tremblay',
-    pronouns: 'he/him',
-    position: 'Impact Intern',
-    major: 'Electrical Engineering Major',
     year: 2025,
-    img: '/images/officers/NoImage.jpeg',
-    alt: '',
-  },
-  {
-    name: 'Ashley Kim',
-    pronouns: 'she/her',
-    position: 'Impact Intern',
-    major: 'Applied Math Major',
-    year: 2023,
-    img: '/images/officers/NoImage.jpeg',
-    alt: '',
+    img: '/images/officers/ElenaYu.jpg',
+    alt: 'a photo of jason smiling in front of some plants',
   },
 ];
 
