@@ -1,19 +1,19 @@
 import Board from './board';
 import Carousel from './carousel';
 import Committees from './committees';
+import Friends from './friends';
 import Internship from './internship';
 import Leadership from './leadership';
-import News from './news';
 import Sponsors from './sponsors';
 import TGProjects from './tgprojects';
 
 export default {
-	committees: Committees,
-	carousel: Carousel,
-	sponsors: Sponsors,
-	leadership: Leadership,
-	news: News,
-	tgprojects: TGProjects,
-	internship: Internship,
-	board: Board,
+  committees: Committees,
+  carousel: Carousel,
+  sponsors: Sponsors,
+  leadership: Leadership,
+  tgprojects: TGProjects,
+  friends: Friends,
+  internship: Internship,
+  board: Board,
 };
