@@ -1,23 +1,31 @@
 import moment from 'moment';
 
 const getCssStringFromCommittee = (committee) => {
-  switch (committee.trim()) {
-    case 'General':
-    case 'Impact': // we should change the branding here soon
-    case 'External':
+  if (!committee) {
+    console.warn('Skipping event. Committee is missing or null.');
+    return null;
+  }
+  switch (committee.toLowerCase().trim()) {
+    case 'general':
+    case 'impact': // we should change the branding here soon
+    case 'external':
+    case 'board':
       return 'board';
-    case 'Teach LA':
+    case 'teach la':
+    case 'teachla':
       return 'teach-la';
-    case 'AI':
-    case 'Cyber':
-    case 'Design':
-    case 'Hack':
-    case 'ICPC':
-    case 'Studio':
-    case 'W':
+    case 'ai':
+    case 'cyber':
+    case 'design':
+    case 'hack':
+    case 'icpc':
+    case 'studio':
+    case 'w':
+    case 'cloud':
       return committee.toLowerCase();
     default:
-      throw new Error(`Unrecognized string ${committee}`);
+      console.warn(`Skipping event. Unrecognized committee: ${committee}`);
+      return null;
   }
 };
 
@@ -32,30 +40,47 @@ const generateSingleEvent = ({
   description,
   links,
   // non-final keys that need to be processed
-  time,
+  rawStart,
+  rawEnd,
   date,
   fblink,
+  image,
 }) => {
-  // TODO: change this logic to actually look at relevant info
-  const allDay = false;
+  let allDay = false;
+
+  // If committee is already null (from getCssStringFromCommittee), skip
+  if (committee === null) {
+    console.warn('Skipping event because committee returned null.');
+    return null;
+  }
 
   if (!start && !end) {
-    if (!date || !time) {
-      throw new Error('Missing either date or time; can\'t proceed');
+    if (!date) {
+      console.warn('Skipping event because date is missing.');
+      return null;
+    }
+    // If rawStart or rawEnd is missing, set allDay to true
+    if (!rawStart) {
+      allDay = true;
+      start = moment(`${date}`, 'YYYY-MM-DD LT').valueOf();
+    } else {
+      const startHr = rawStart.trim();
+      start = moment(`${date} ${startHr}`, 'YYYY-MM-DD LT').valueOf();
     }
 
-    // const [year, month, day] = date.split('-');
-    const [rawStart, rawEnd] = time.split('-');
-    const startHr = rawStart.trim();
-    const endHr = rawEnd.trim().split(' ')[0];
-    start = moment(`${date} ${startHr} PM`, 'YYYY-MM-DD LT').valueOf();
-    end = moment(`${date} ${endHr} PM`, 'YYYY-MM-DD LT').valueOf();
+    if (!rawEnd) {
+      allDay = true;
+      end = moment(`${date}`, 'YYYY-MM-DD LT').valueOf();
+    } else {
+      const endHr = rawEnd.trim();
+      end = moment(`${date} ${endHr}`, 'YYYY-MM-DD LT').valueOf();
+    }
   }
 
-  if(!links){
+  if (!links) {
     links = [];
   }
-  if(location.includes('ucla.zoom.us')) {
+  if (location.includes('ucla.zoom.us')) {
     const zoomLink = location;
     location = 'Zoom';
     links.push({
@@ -66,10 +91,38 @@ const generateSingleEvent = ({
   }
   if (fblink) {
     links.push({
-      text: 'Facebook Event',
+      text: 'Link',
       href: fblink,
       ext: true,
     });
+  }
+  if (image) {
+    try {
+      // Extract host from url to comply with codeQL scanner
+      const url = new URL(image);
+      const host = url.hostname;
+
+      if (host === 'drive.google.com') {
+        // Google drive urls need to be transformed
+        // 'https://drive.google.com/uc?export=view&id=' + image id (id which follows /d in original link)
+        image =
+          'https://drive.google.com/uc?export=view&id=' +
+          image.substring(image.indexOf('/d/') + 3, image.indexOf('/view'));
+      } else {
+        image = url.href;
+      }
+    } catch (e) {
+      console.log('Warning: Invalid Image URL: ' + e);
+      image = '/images/events/default-event.png';
+    }
+  } else {
+    // Default image
+    image = '/images/events/default-event.png';
+  }
+
+  if (!title) {
+    console.warn('Skipping event because title is missing.');
+    return null;
   }
 
   return {
@@ -82,6 +135,7 @@ const generateSingleEvent = ({
     committee,
     description,
     links,
+    image,
   };
 };
 
